Optimize vSphere convert_vm_mob_ref_to_attr_hash
This patch reduces the number of round trips to the vSphere API by
using the `collect!` method on the ManagedObject, vm_mob_ref, which
retrieves most of the properties in one request.

For the remaining properties: hypervisor and mac_addresses, we still
need to make additional requests.

Overall this patch provides a nice speed improvement for the
`convert_vm_mob_ref_to_attr_hash` method.