(#?????) WIP - Make fog fast and return simple data in the models
The goal of this work in progress is to preserver Jeff's work making the
models simple while also preserving Carl's work speeding things up.

The focal point is in the convert_vm_mob_ref_to_attr_hash method which
needs to return a simple hash using only strings and arrays and such.
No rbvmomi objects.

This hash will be used by Fog to populate the server model, so the keys
needs to match up with the attributes.

We've replaced the hypervisor attribute with the name string and are in
the process of getting rid of the host alias.